Simplify object naming
* Remove `name` parameter
* Use <Plug>Wobble_ + default mapping as name

This makes the names a bit more opaque, but should make the documentation clearer, by not requiring so many cross-references. Now, we can refer to the objects by default mapping everywhere, and give clear, simple instructions for how to derive the name, rather than needing to list out a table of every name and default mapping.